Move logic into the model
