+++ administrator/components/com_media/controllers/api.json.php
@@ -9,6 +9,8 @@
 
 defined('_JEXEC') or die;
 
+use Joomla\Cms\Controller\Controller;
+
 /**
  * Api Media Controller
  *
@@ -16,41 +18,8 @@ defined('_JEXEC') or die;
  *
  * @since  __DEPLOY_VERSION__
  */
-class MediaControllerApi extends JControllerLegacy
+class MediaControllerApi extends Controller
 {
-	/**
-	 * The local file adapter to work with.
-	 *
-	 * @var MediaFileAdapterInterface
-	 */
-	protected $adapter = null;
-
-	/**
-	 * Constructor.
-	 *
-	 * @param   array  $config  An optional associative array of configuration settings.
-	 * Recognized key values include 'name', 'default_task', 'model_path', and
-	 * 'view_path' (this list is not meant to be comprehensive).
-	 *
-	 * @since   3.0
-	 */
-	public function __construct($config = array())
-	{
-		parent::__construct($config);
-
-		if (!isset($config['fileadapter']))
-		{
-			// Compile the root path
-			$root = JPATH_ROOT . '/' . JComponentHelper::getParams('com_media')->get('file_path', 'images');
-			$root = rtrim($root) . '/';
-
-			// Default to the local adapter
-			$config['fileadapter'] = new MediaFileAdapterLocal($root);
-		}
-
-		$this->adapter = $config['fileadapter'];
-	}
-
 	/**
 	 * Api endpoint for the media manager front end. The HTTP methods GET, PUT, POST and DELETE
 	 * are supported.
@@ -115,47 +84,50 @@ class MediaControllerApi extends JControllerLegacy
 		// Determine the method
 		$method = $this->input->getMethod() ? : 'GET';
 
+		// Check token for requests which do modify files
+		if (in_array(strtolower($method), array('post', 'put', 'delete')) && !JSession::checkToken('request'))
+		{
+			$this->sendResponse(new Exception(JText::_('JINVALID_TOKEN'), 403));
+			return;
+		}
+
 		try
 		{
-			// Gather the data accoring to the method
+			// Gather the data according to the method
 			switch (strtolower($method))
 			{
 				case 'get':
-					$data = $this->adapter->getFiles($path, $this->input->getWord('filter'));
+					$data = $this->getModel()->getFiles($path, $this->input->getWord('filter'));
 					break;
 				case 'delete':
-					$this->adapter->delete($path);
+					$this->getModel()->delete($path);
 					break;
 				case 'post':
 					$content      = $this->input->json;
-					$name         = $this->getSafeName($content->get('name'));
+					$name         = $content->get('name');
 					$mediaContent = base64_decode($content->get('content'));
 
 					if ($mediaContent)
 					{
-						$this->checkContent($name, $mediaContent);
-
 						// A file needs to be created
-						$this->adapter->createFile($name, $path, $mediaContent);
+						$this->getModel()->createFile($name, $path, $mediaContent);
 					}
 					else
 					{
 						// A file needs to be created
-						$this->adapter->createFolder($name, $path);
+						$this->getModel()->createFolder($name, $path);
 					}
 
-					$data = $this->adapter->getFile($path . '/' . $name);
+					$data = $this->getModel()->getFile($path . '/' . $name);
 					break;
 				case 'put':
 					$content      = $this->input->json;
 					$name         = basename($path);
 					$mediaContent = base64_decode($content->get('content'));
 
-					$this->checkContent($name, $mediaContent);
+					$this->getModel()->updateFile($name, str_replace($name, '', $path), $mediaContent);
 
-					$this->adapter->updateFile($name, str_replace($name, '', $path), $mediaContent);
-
-					$data = $this->adapter->getFile($path . '/' . $name);
+					$data = $this->getModel()->getFile($path . '/' . $name);
 					break;
 				default:
 					throw new BadMethodCallException('Method not supported yet!');
@@ -204,78 +176,19 @@ class MediaControllerApi extends JControllerLegacy
 		echo new JResponseJson($data);
 	}
 
-
 	/**
-	 * Creates a safe file name for the given name.
+	 * Method to get a model object, loading it if required.
 	 *
-	 * @param   string  $name  The filename
+	 * @param   string  $name    The model name. Optional.
+	 * @param   string  $prefix  The class prefix. Optional.
+	 * @param   array   $config  Configuration array for model. Optional.
 	 *
-	 * @return  string
+	 * @return  Model|boolean  Model object on success; otherwise false on failure.
 	 *
-	 * @since   __DEPLOY_VERSION__
-	 * @throws  Exception
+	 * @since   3.0
 	 */
-	private function getSafeName($name)
+	public function getModel($name = 'Api', $prefix = 'MediaModel', $config = array())
 	{
-		// Make the filename safe
-		$name = JFile::makeSafe($name);
-
-		// Transform filename to punycode
-		$name = JStringPunycode::toPunycode($name);
-
-		$extension = JFile::getExt($name);
-
-		if ($extension)
-		{
-			$extension = '.' . strtolower($extension);
-		}
-
-		// Transform filename to punycode, then neglect other than non-alphanumeric characters & underscores.
-		// Also transform extension to lowercase.
-		$name = preg_replace(array("/[\\s]/", '/[^a-zA-Z0-9_]/'), array('_', ''), $name) . $extension;
-
-		return $name;
-	}
-
-	/**
-	 * Performs various check if it is allowed to save the content with the given name.
-	 *
-	 * @param   string  $name          The filename
-	 * @param   string  $mediaContent  The media content
-	 *
-	 * @return  void
-	 *
-	 * @since   __DEPLOY_VERSION__
-	 * @throws  Exception
-	 */
-	private function checkContent($name, $mediaContent)
-	{
-		if (!JFactory::getUser()->authorise('core.create', 'com_media'))
-		{
-			throw new Exception(JText::_('COM_MEDIA_ERROR_CREATE_NOT_PERMITTED'), 403);
-		}
-
-		$helper = new JHelperMedia();
-		$serverlength = $this->input->server->get('CONTENT_LENGTH');
-		if ($serverlength > ($params->get('upload_maxsize', 0) * 1024 * 1024)
-			|| $serverlength > $helper->toBytes(ini_get('upload_max_filesize'))
-			|| $serverlength > $helper->toBytes(ini_get('post_max_size'))
-			|| $serverlength > $helper->toBytes(ini_get('memory_limit')))
-		{
-			throw new Exception(JText::_('COM_MEDIA_ERROR_WARNFILETOOLARGE'));
-		}
-
-		// @todo find a better way to check the input, by not writing the file to the disk
-		$tmpFile = JFactory::getApplication()->getConfig()->get('tmp_path') . '/' . uniqid($name);
-
-		if (!JFile::write($tmpFile, $mediaContent))
-		{
-			throw new Exception(JText::_('JLIB_MEDIA_ERROR_UPLOAD_INPUT'));
-		}
-
-		if (!$helper->canUpload(array('name' => $name, 'size' => sizeof($mediaContent), 'tmp_name' => $tmpFile), 'com_media'))
-		{
-			throw new Exception(JText::_('COM_MEDIA_ERROR_UNABLE_TO_UPLOAD_FILE'), 403);
-		}
+		return parent::getModel($name, $prefix, $config);
 	}
 }
